requirements.txt
beautifulsoup4
requests
python-whois
dnspython
tldextract
pdfkit
python-nmap
